Fix database id types

+++ src/main/resources/db/migration/V3__create_comments_table.sql
@@ -2,7 +2,7 @@ DROP TABLE IF EXISTS comments;
 
 CREATE TABLE comments (
   id bigserial PRIMARY KEY,
-  user_id INT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
-  post_id INT NOT NULL REFERENCES posts(id) ON DELETE CASCADE,
+  user_id BIGINT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+  post_id BIGINT NOT NULL REFERENCES posts(id) ON DELETE CASCADE,
   content varchar(250) NOT NULL
 );
